requirements.txt
screen-brightness-control==0.14.0
requests==2.31.0
astral==3.2
pystray==0.19.4
Pillow==9.5.0
